feat(workflows): abstract debug trait
