+++ packages/common/chirp-workflow/core/tests/common.rs
@@ -118,7 +118,7 @@ pub async fn setup_dependencies(fdb: bool) {
 			tokio::spawn(start_fdb());
 			create_fdb_db().await;
 		} else {
-			tokio::time::sleep(std::time::Duration::from_millis(500)).await;
+			tokio::time::sleep(std::time::Duration::from_secs(1)).await;
 		}
 	}
 }
